Make the thrust detent bindings respect selected engines
Now that PgUp/PgDn respect selected engines respects selected engines,
make the TOGA, MCT/FLX and CL shortcuts do so as well.

Idle binding is an exception, because both thrust levers must be set to
idle for spoilers to deploy while one is enough to disengage the A/THR
and you don't want to repeat TAM3054 just because you forgot the
re-select both engines.