Make reverse thrust also respect selected engines
The simple toggle, bound to Del, engages reverse on the selected engine,
but if it is engaged on either, disengages it on both, so you don't get
stuck in reverse when selecting engine.

I also removed the check of weight-on-wheels from the disengage case, so
you don't get stuck if somehow the aircraft still becomes airborne after
selecting reverse on the ground.

The finer control, bound to F1 and F2, only controls the selected engines
now. Previous version disengaged reverse for both engines when either
reached zero throttle-rev, which does not seem to be founded in reality
and does not make much sense. You can always use Del to disengage reverse
on both engines.